% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indexARIA}
\alias{indexARIA}
\title{Indexed values of 96 cells.}
\format{A data frame with 96 rows and 8 variables:
\describe{
  \item{IdxRow}{indexed row}
  \item{IdxCol}{indexed column}
  \item{FSC}{forward scatter}
  \item{SSC}{side scatter}
  \item{APC}{allophycocyanin}
  \item{FITC}{fluorescein}
  \item{PerCP}{peridinin-chlorophyll}
  \item{PI}{propidium iodide}
  ...
}}
\source{
personal work
}
\usage{
indexARIA
}
\description{
A dataset containing the indexed values of 96 single-cells.
}
\keyword{datasets}
