% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputARIA}
\alias{inputARIA}
\title{Values of 96 cells.}
\format{A flowFrame with 96 rows and 8 variables:
\describe{
  \item{FSC}{forward scatter}
  \item{SSC}{side scatter}
  \item{APC}{allophycocyanin}
  \item{FITC}{fluorescein}
  \item{PerCP}{peridinin-chlorophyll}
  \item{PI}{propidium iodide}
  \item{$`INDEX SORTING LOCATIONS_x`}{location keyword}
  \item{$`$CYT`}{cytometer keyword}
  ...
}}
\source{
Handcrafted, personal work
}
\usage{
inputARIA
}
\description{
A flowFrame containing the values of 96 single-cells and
location keywords.
}
\keyword{datasets}
