% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_index.R
\name{retrieve_index}
\alias{retrieve_index}
\title{Retrieve Index Sorted Data for a Given Cell Sorter}
\usage{
retrieve_index(data, sorter = "aria", ...)
}
\arguments{
\item{data}{A index sorted .fcs file}

\item{sorter}{A given cell sorter}

\item{...}{Extra arguments, not used}
}
\description{
This function retrieves index sorted data from various cell sorters.
}
\examples{

}
