% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_plate.R
\name{explore_plate}
\alias{explore_plate}
\title{Plot a Plate Visual from Index Sorted Data}
\usage{
explore_plate(
  data,
  vars = c("IdxRow", "IdxCol"),
  cell = TRUE,
  dim = c(8, 12),
  legend = TRUE,
  circle = 12,
  ...
)
}
\arguments{
\item{data}{A index sorted .fcs file}

\item{vars}{Features to be used}

\item{cell}{Display cell number values}

\item{dim}{A plate's dimension c(x, y)}

\item{legend}{Display fluorescence legend}

\item{circle}{Change wells' size}

\item{...}{Extra arguments, not used}
}
\description{
This function plots a plate visual from index sorted data.
}
\examples{

}
