% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexSort.R
\name{explore_plate}
\alias{explore_plate}
\title{Plot a Plate Visual from Index Sorted Data}
\usage{
explore_plate(
  data,
  param = c("APC", "FITC"),
  cell = TRUE,
  dim = c(8, 12),
  legend = TRUE,
  circle = 12,
  ...
)
}
\arguments{
\item{data}{An indexed `data.frame` object}

\item{param}{Features to be displayed as MFI}

\item{cell}{Display or not cell number values}

\item{dim}{A plate's dimension c(x, y)}

\item{legend}{Display or not fluorescence legend}

\item{circle}{Size of the wells on illustration}

\item{...}{Extra arguments, not used}
}
\value{
A `ggplot` object
}
\description{
The \code{explore_plate} is used to generate a set of 
illustrations that automatically display the Median Fluorescent 
Intensity (MFI) per well for the selected parameters. Additionally, 
the number of cells will be displayed for each well.
}
\examples{
library(tidyverse)

explore_plate(indexARIA, param = c("APC", "FITC", "PerCP", "PI"))

# Remove the legend
explore_plate(indexARIA, param = c("APC", "FITC", "PerCP", "PI"), legend = FALSE)

}
