% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_index.R
\name{retrieve_index}
\alias{retrieve_index}
\title{Retrieve Index Sorted Data for a Given Cell Sorter}
\usage{
retrieve_index(data, sorter = "aria", ...)
}
\arguments{
\item{data}{An index sorted .fcs file}

\item{sorter}{Specify the cell sorter. Possible values are `aria`,
`symphony`, `influx`, `jazz`, `astrios`.}

\item{...}{Extra arguments, not used}
}
\value{
A `data.frame` object
}
\description{
The `retrieve_index` was designed in order to give 
researchers the ability to automatically gate wells from .fcs data 
generated using various cell sorters. It will generate parameters 
corresponding to the `x` and `y` axes (IdxRow and IdxCol).
}
\examples{
library(flowCore)

inputFCS <- read.FCS("~/path/to/file.fcs")

# Retrieve indexed data from an Aria
result <- retrieve_index(inputFCS)

# Retrieve data from another sorter
result <- retrieve_index(inputFCS, sorter = "influx")

}
