% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexSort.R
\name{retrieve_index}
\alias{retrieve_index}
\title{Retrieve Index Sorted Data for a Given Cell Sorter}
\usage{
retrieve_index(data, sorter = "auto", ...)
}
\arguments{
\item{data}{An index sorted .fcs file}

\item{sorter}{Specify the cell sorter. Possible values are `aria`,
`symphony`, `influx`, `jazz`, `astrios`. If `auto` (default) will  
 use a regular expression to identify the sorter based on the $CYT  
 keyword.}

\item{...}{Extra arguments, not used}
}
\value{
A `data.frame` object
}
\description{
The \code{retrieve_index} was designed in order to give  
researchers the ability to automatically gate wells from .fcs data  
generated using various cell sorters. It will generate parameters  
corresponding to the `x` and `y` axes (IdxRow and IdxCol).
}
\examples{
library(flowCore)
library(tidyverse)

# Retrieve indexed data from an Aria
retrieve_index(inputARIA)

# Retrieve data from specific sorter
result <- retrieve_index(inputARIA, sorter = "aria")

}
